+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/BundledHlsMediaChunkExtractor.java
@@ -17,6 +17,7 @@ package androidx.media3.exoplayer.hls;
 
 import static androidx.media3.common.util.Assertions.checkState;
 
+import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.media3.common.Format;
 import androidx.media3.common.util.TimestampAdjuster;
@@ -27,6 +28,8 @@ import androidx.media3.extractor.ExtractorOutput;
 import androidx.media3.extractor.PositionHolder;
 import androidx.media3.extractor.mp3.Mp3Extractor;
 import androidx.media3.extractor.mp4.FragmentedMp4Extractor;
+import androidx.media3.extractor.text.SubtitleParser;
+import androidx.media3.extractor.text.SubtitleTranscodingExtractor;
 import androidx.media3.extractor.ts.Ac3Extractor;
 import androidx.media3.extractor.ts.Ac4Extractor;
 import androidx.media3.extractor.ts.AdtsExtractor;
@@ -45,6 +48,7 @@ public final class BundledHlsMediaChunkExtractor implements HlsMediaChunkExtract
   @VisibleForTesting /* package */ final Extractor extractor;
   private final Format multivariantPlaylistFormat;
   private final TimestampAdjuster timestampAdjuster;
+  @Nullable private final SubtitleParser.Factory subtitleParserFactory;
 
   /**
    * Creates a new instance.
@@ -55,9 +59,35 @@ public final class BundledHlsMediaChunkExtractor implements HlsMediaChunkExtract
    */
   public BundledHlsMediaChunkExtractor(
       Extractor extractor, Format multivariantPlaylistFormat, TimestampAdjuster timestampAdjuster) {
+    this(
+        extractor,
+        multivariantPlaylistFormat,
+        timestampAdjuster,
+        /* subtitleParserFactory= */ null);
+  }
+
+  /**
+   * Creates a new instance.
+   *
+   * @param extractor The underlying {@link Extractor}.
+   * @param multivariantPlaylistFormat The {@link Format} obtained from the multivariant playlist.
+   * @param timestampAdjuster A {@link TimestampAdjuster} to adjust sample timestamps.
+   * @param subtitleParserFactory A {@link SubtitleParser.Factory} to be used with WebVTT subtitles.
+   *     If the value is null, subtitles will be parsed during decoding, otherwise - during
+   *     extraction. Decoding will only work if this subtitleParserFactory supports the provided
+   *     multivariantPlaylistFormat.
+   */
+  // TODO(b/289983417): Once the subtitle-parsing-during-extraction is the only available flow, make
+  // this constructor public and remove @Nullable from subtitleParserFactory
+  /* package */ BundledHlsMediaChunkExtractor(
+      Extractor extractor,
+      Format multivariantPlaylistFormat,
+      TimestampAdjuster timestampAdjuster,
+      @Nullable SubtitleParser.Factory subtitleParserFactory) {
     this.extractor = extractor;
     this.multivariantPlaylistFormat = multivariantPlaylistFormat;
     this.timestampAdjuster = timestampAdjuster;
+    this.subtitleParserFactory = subtitleParserFactory;
   }
 
   @Override
@@ -97,6 +127,11 @@ public final class BundledHlsMediaChunkExtractor implements HlsMediaChunkExtract
     if (extractor instanceof WebvttExtractor) {
       newExtractorInstance =
           new WebvttExtractor(multivariantPlaylistFormat.language, timestampAdjuster);
+      if (subtitleParserFactory != null
+          && subtitleParserFactory.supportsFormat(multivariantPlaylistFormat)) {
+        newExtractorInstance =
+            new SubtitleTranscodingExtractor(newExtractorInstance, subtitleParserFactory);
+      }
     } else if (extractor instanceof AdtsExtractor) {
       newExtractorInstance = new AdtsExtractor();
     } else if (extractor instanceof Ac3Extractor) {
@@ -110,7 +145,7 @@ public final class BundledHlsMediaChunkExtractor implements HlsMediaChunkExtract
           "Unexpected extractor type for recreation: " + extractor.getClass().getSimpleName());
     }
     return new BundledHlsMediaChunkExtractor(
-        newExtractorInstance, multivariantPlaylistFormat, timestampAdjuster);
+        newExtractorInstance, multivariantPlaylistFormat, timestampAdjuster, subtitleParserFactory);
   }
 
   @Override