+++ libraries/exoplayer_hls/src/test/java/androidx/media3/exoplayer/hls/HlsMediaPeriodTest.java
@@ -97,7 +97,8 @@ public final class HlsMediaPeriodTest {
               HlsMediaSource.METADATA_TYPE_ID3,
               /* useSessionKeys= */ false,
               PlayerId.UNSET,
-              /* timestampAdjusterInitializationTimeoutMs= */ 0);
+              /* timestampAdjusterInitializationTimeoutMs= */ 0,
+              /* subtitleParserFactory= */ null);
         };
 
     MediaPeriodAsserts.assertGetStreamKeysAndManifestFilterIntegration(